requirements.txt
ccxt>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
schedule>=1.2.0
flask>=3.0.0
